services/datalad/pyproject.toml
[project]
name = "datalad-service"
description = "HTTP interfaces for manipulating DataLad datasets"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aioshutil>=1.5",
    "bidsschematools>=1.0.10",
    "boto3>=1.38.30",
    "charset-normalizer>=3.4.2",
    "dnspython>=2.7.0",
    "falcon>=4.0.2",
    "pygit2>=1.18.0",
    "pygithub>=2.6.1",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
    "sentry-sdk[falcon]>=2.29.1",
    "taskiq[reload]>=0.11.18",
    "taskiq-redis>=1.0.9",
    "uvicorn[standard]>=0.34.3",
    "httpx>=0.28.1",
]
dynamic = ["version"]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "coverage>=7.8.2",
    "datalad>=1.2.0",
    "exceptiongroup>=1.3.0",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "pytest-xdist>=3.7.0",
    "ruff>=0.12.2",
]

[tool.hatch.version]
source = "vcs"
fallback-version = "0+unknown"

[tool.hatch.version.raw-options]
root = "../.."

[tool.ruff.format]
quote-style = "single"
